+++ core-bundle/src/Resources/contao/themes/flexible/src/hover.js
@@ -189,6 +189,27 @@ var Theme = {
 	},
 
 	/**
+	 * Set up the profile toggle
+	 */
+	setupProfileToggle: function() {
+		var tmenu = $('tmenu'),
+			ul = tmenu.getElement('.level_2'),
+			h2 = tmenu.getElement('h2');
+		if (!ul || !h2) return;
+
+		h2.addEvent('click', function(e) {
+			ul.fade();
+			e.stopPropagation();
+		});
+
+		$(document.body).addEvent('click', function() {
+			if (ul.getStyle('opacity')) {
+				ul.fade('out');
+			}
+		});
+	},
+
+	/**
 	 * Hide the menu on scroll
 	 */
 	hideMenuOnScroll: function() {
@@ -276,6 +297,7 @@ window.addEvent('domready', function() {
 	Theme.setupCtrlClick();
 	Theme.setupTextareaResizing();
 	Theme.setupMenuToggle();
+	Theme.setupProfileToggle();
 	Theme.hideMenuOnScroll();
 	Theme.setupSplitButtonToggle();
 });